# functions/openapi.yaml
openapi: 3.1.0
info:
  title: LumiMD API
  version: 1.0.0
  description: |
    Backend API for LumiMD - a patient-facing medical navigation app.
    
    ## Authentication
    Most endpoints require a Firebase ID token in the Authorization header:
    ```
    Authorization: Bearer <firebase-id-token>
    ```
    
    ## Error Format
    All errors follow this format:
    ```json
    {
      "code": "error_code",
      "message": "Human readable message",
      "details": { "field": "optional context" }
    }
    ```
servers:
  - url: https://us-central1-lumimd.cloudfunctions.net/api
    description: Production
  - url: https://us-central1-lumimd-dev.cloudfunctions.net/api
    description: Development
...
paths:
...
  /v1/auth/create-handoff:
    post:
      tags: [auth]
      summary: Create auth handoff code
      description: |
        Creates a one-time code for seamless mobile → web authentication.
        Used when mobile app opens web portal to avoid re-login.
        
        Code expires in 5 minutes and can only be used once.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Handoff code created successfully
          content:
            application/json:
              schema:
                type: object
                required: [code]
                properties:
                  code:
                    type: string
                    description: One-time handoff code (valid for 5 minutes)
                    example: "xyz789abc123..."
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /v1/auth/exchange-handoff:
    post:
      tags: [auth]
      summary: Exchange handoff code for custom token
      description: |
        Exchanges a one-time handoff code for a Firebase custom token.
        Web portal uses this to authenticate users coming from mobile.
        
        The code is invalidated after use to prevent replay attacks.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  description: The handoff code from create-handoff
                  example: "xyz789abc123..."
      responses:
        '200':
          description: Custom token generated successfully
          content:
            application/json:
              schema:
                type: object
                required: [token]
                properties:
                  token:
                    type: string
                    description: Firebase custom token for signInWithCustomToken()
                    example: "eyJhbGc..."
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid, expired, or already-used code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'

  # ==================== Visits ====================
...
  /v1/actions:
    get:
      tags: [actions]
      summary: List action items
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, done]
      responses:
        '200':
          description: List of actions
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ActionItem'

    post:
      tags: [actions]
      summary: Create action item
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionItemInput'
      responses:
        '201':
          description: Action created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionItem'
  /v1/actions/{id}:
    patch:
      tags: [actions]
      summary: Update action item
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionItemPatch'
      responses:
        '200':
          description: Action updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionItem'

    delete:
      tags: [actions]
      summary: Delete action item
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Action deleted

  # ==================== Medications ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [unauthorized, forbidden, not_found, validation_failed, rate_limited, conflict, server_error]
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    ActionItem:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        visitId:
          type: string
        title:
          type: string
        subtitle:
          type: string
        due:
          type: string
          format: date
        status:
          type: string
          enum: [open, done]
        critical:
          type: boolean
    ActionItemInput:
      type: object
      required: [title, subtitle]
      properties:
        title:
          type: string
          example: "Schedule follow-up"
        subtitle:
          type: string
          example: "Cardiology – 2 weeks"
        due:
          type: string
          format: date
          example: "2024-02-01"
        critical:
          type: boolean
          default: false
    ActionItemPatch:
      type: object
      properties:
        title:
          type: string
        subtitle:
          type: string
        due:
          type: string
          format: date
        status:
          type: string
          enum: [open, done]
        critical:
          type: boolean
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "unauthorized"
            message: "Missing or invalid authorization header"
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "validation_failed"
            message: "Invalid request body"
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "server_error"
            message: "An unexpected error occurred"